+++ 05-ScenceManager/FlyingWood.h
@@ -10,7 +10,7 @@ class FlyingWood : public CGameObject
 {
 public:
 	bool IsCollMario;
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* colliable_objects = NULL);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* colliable_objects = NULL);
 	virtual void Render();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 };
