+++ 05-ScenceManager/GameObject.h
@@ -61,7 +61,7 @@ public:
 	int state;
 	bool BBox = true;
 	bool checkOnCam;
-	DWORD dt; 
+	ULONGLONG dt; 
 
 	LPANIMATION_SET animation_set;
 
@@ -94,7 +94,7 @@ public:
 	CGameObject();
 
 	virtual void GetBoundingBox(float &left, float &top, float &right, float &bottom) = 0;
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT> *coObjects = NULL);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT> *coObjects = NULL);
 	virtual void Render() = 0;
 	virtual void SetState(int state) { this->state = state; }
