+++ 05-ScenceManager/Line.h
@@ -10,7 +10,7 @@ public:
 	Line(float Width);
 	virtual void Render();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
-	int GetX_Left() { return X_Left; };
+	float GetX_Left() { return X_Left; };
 	float GetX_Right() { return X_Right; };
 };
