+++ 05-ScenceManager/MusicNote.h
@@ -38,7 +38,7 @@ public:
 	float vxOfPlayer = 0;
 	MusicNote(int typeNote);
 	virtual void GetBoundingBox(float& left, float& top, float& right, float& bottom);
-	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects = NULL);
+	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects = NULL);
 	virtual void Render();
 	virtual void SetState(int state);
 };
