+++ 05-ScenceManager/ObjCheckFalling.h
@@ -6,7 +6,7 @@ class ObjCheckFalling :public CGameObject
 public:
 	bool isFalling = false;
 	ObjCheckFalling();
-	void Update(DWORD dt, vector<LPGAMEOBJECT>* listMapObj);
+	void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* listMapObj);
 	void Render();
 	void GetBoundingBox(float& l, float& t, float& r, float& b);
 };
